Conflicts:
	.travis.yml
	doc/doxygen/install/install-win.doxygen
	src/openms/source/SYSTEM/SysInfo.cpp

+++ src/openms/include/OpenMS/FORMAT/Base64.h
@@ -237,7 +237,7 @@ private:
         for (Size i = 0; i < in.size(); ++i)
         {
           Reinterpreter64_ tmp;
-          tmp.f = in[i];
+          tmp.f = static_cast<double>(in[i]);
           tmp.i = endianize64(tmp.i);
           in[i] = tmp.f;
         }
